+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Unreleased
 
+- Implement `MonitorId::get_dimensions` for Android.
+- Fixed windows not receiving mouse events when click-dragging the mouse outside the client area of a window, on Windows platforms.
 - Added `set_min_dimensions` and `set_max_dimensions` methods to `Window`, and implemented on Windows, X11, Wayland, and OSX.
+
 
 # Version 0.11.0 (2018-02-09)